+++ src/bwgraph/BipWGSS.java
@@ -28,12 +28,12 @@ public interface BipWGSS {
 	
 	// Returns weight of edge from i to j
 	// Returns 0 if no such edge exists
-	int getWeight(int l, int r);
+	double getWeight(int l, int r);
 	
 	// Adds directed edge from vertex l to vertex r with weight w
 	// Overrides if there is already an edge
 	// Does nothing if l and r are in the same set
-	void addEdge(int l, int r, int w);
+	void addEdge(int l, int r, double w);
 	
 	// Returns True if there is an edge from l to r, False otherwise
 	boolean edgeQ(int l, int r);
@@ -44,23 +44,23 @@ public interface BipWGSS {
 	// Receives a vertex v and an array of distances dist
 	// Checks if there is a path from u to sink such that dist along that path never decreases
 	// If there is, returns 'true' and inverts it. Otherwise, returns 'false'
-	boolean invertPathBF(int v, int[] dist);
+	boolean invertPathBF(int v, double[] dist);
 	
 	// Finds a path from source to sink through Bellman-Ford to minimize the cost
 	// Returns the list of distances to the source
-	int[] BellmanFord();
+	double[] BellmanFord();
 	
 	// Receives a vertex v, an array of distances dist and an array of function values h
 	// Checks if there is a path from u to sink such that dist along that path never decreases
 	// If there is, returns 'true' and inverts it. Otherwise, returns 'false'
-	boolean invertPathDijkstra(int v, int[] dist, int[] h);
+	boolean invertPathDijkstra(int v, double[] dist, double[] h);
 	
 	// Finds a path from source to sink through Dijkstra to minimize the cost
 	// Returns the list of distances to the source
-	int[] Dijkstra(int[] h);
+	double[] Dijkstra(double[] h);
 	
 	// Finds a path from source to sink through Dijkstra to minimize the cost using Priority Queue
 	// Returns the list of distances to the source
-	int[] DijkstraPQ(int[] h);
+	double[] DijkstraPQ(double[] h);
 		
 }
